backend/requirements.txt
Django==5.1.4
djangorestframework==3.14.0
mongoengine==0.27.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
asgiref==3.8.1
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.1
PyJWT==2.10.1
requests==2.32.3
sqlparse==0.5.3
twilio==9.4.3
tzdata==2024.2
urllib3==2.3.0
yarl==1.18.3
